+++ app/Http/Controllers/passwordController.php
@@ -3,9 +3,9 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Helpers\Token;
 use App\Category;
 use App\Password;
+use App\User;
 
 class passwordController extends Controller
 {
@@ -39,15 +39,37 @@ class passwordController extends Controller
     {
         $password = new Password();
 
-        $category_name = $request->category_name;
-        $category = Category::where('name', $category_name)->first();
-        $password = Password::where('name', $category_name)->first();
-        
-        $password->add_password($request, $category);
+        $email = $request->data_token->email;
+        $user = User::where('email', $email)->first();
 
-        return response()->json([
-            "message" => "contraseña creada"
-        ], 200);
+        $category_name = $request->category_name;
+        $password_title = $request->title;
+        $category = Category::where('user_id', $user->id)->where('name', $category_name)->first();
+
+        if(!isset($category))
+        {
+            return response()->json([
+                "message" => "La categoria no existe"
+            ], 401);
+        }
+        else
+        {
+            $password_Searched = Password::where('category_id', $category->id)->where('title', $password_title)->first();
+            if(!isset($password_Searched))
+            {
+                $password->add_password($request, $category->id);
+
+                return response()->json([
+                    "message" => "contraseña creada"
+                ], 200);
+            }
+            else
+            {
+                return response()->json([
+                    "message" => "La contraseña ya existe"
+                ], 401);
+            }
+        }
     }
 
     /**
@@ -56,9 +78,32 @@ class passwordController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Request $request)
     {
-        //
+        $email = $request->data_token->email;
+        $user = User::where('email', $email)->first();
+
+        $category_name = $request->name_category;
+        $category = Category::where('name', $category_name)->first();
+       // var_dump($category);exit;
+       if ($category->user_id == $user->id)
+       {
+        $password = Password::where('category_id', $category->id)->get();
+
+        foreach ($password as $key => $value)
+        {
+            return response()->json([
+                "Passwords" => $password,
+            ]);
+        }
+       }
+       else
+       {
+        return response()->json([
+            "message" => "no existe esa categoria"
+        ]);
+       }
+        
     }
 
     /**
@@ -81,12 +126,34 @@ class passwordController extends Controller
      */
     public function update(Request $request)
     {
-        $password_LastName = $request->LastName;
-        $password = Password::where('title', $password_LastName)->first();
+        $email = $request->data_token->email;
+        $user = User::where('email', $email)->first();
 
-        $password->title = $request->NewName;
-        $password->password = $request->Newpassword;
-        $password->update();
+        $password_LastName = $request->LastName;
+        $categoryName = $request->categoryName;
+        $category = Category::where('user_id',$user->id)->where('name',$categoryName)->first();
+        
+        if (isset($category)) {
+            $password = Password::where('title', $password_LastName)->where('category_id', $category->id)->first();
+           if (!isset($password)) {
+                 return response()->json([
+                     "Error" => "No existe la contraseña"
+                    ], 401);
+            }else{                
+                
+                $password->title = $request->NewName;
+                $password->password = $request->Newpassword;
+                //var_dump($password->title);exit;
+                $password->update();
+                return response()->json([
+                    "message" => "Se ha modificado la contraseña"
+                ], 201);
+            }
+        }else{
+             return response()->json([
+                 "message" => "No existe la categoria"
+                ], 401);
+        }
     }
 
     /**
@@ -97,13 +164,30 @@ class passwordController extends Controller
      */
     public function destroy(Request $request)
     {
+        $email = $request->data_token->email;
+        $user = User::where('email', $email)->first();
+
         $password_title = $request->name;
-        $password = Password::where('title', $password_title)->first();
+        $categoryName = $request->categoryName;
+        $category = Category::where('user_id',$user->id)->where('name',$categoryName)->first();
 
-        $password->delete();
-
+        if (!isset($category)) {
             return response()->json([
-                "message" => 'la contrasena ha sido eliminado'
-            ], 200);
+                "message" => "No existe la categoria"
+            ], 401);
+        }else
+        {
+            $password = Password::where('title', $password_title)->where('category_id', $category->id)->first();
+            if (!isset($password)) {
+                return response()->json([
+                    "message" => "No existe la contraseña"
+                ], 401);
+            }else{                
+                $password->delete();
+                return response()->json([
+                    "message" => "La contraseña ha sido eliminada"
+                ], 200);
+        }
     }
+}
 }
